Seperate the custom card creator into its own file
